add tests + separate package namespaces
