+++ build.gradle
@@ -39,6 +39,7 @@ dependencies {
             project(':fahrschein-http-jdk11'),
             project(':fahrschein-inmemory'),
             project(':fahrschein-metrics-dropwizard'),
+            project(':fahrschein-metrics-micrometer'),
             project(':fahrschein-typeresolver'),
             project(':fahrschein-opentracing'),
             project(':fahrschein-opentelemetry'),
