+++ fahrschein-metrics-dropwizard/src/main/java/org/zalando/fahrschein/metrics/dropwizard/DropwizardMetricsCollector.java
@@ -1,4 +1,4 @@
-package org.zalando.fahrschein.metrics;
+package org.zalando.fahrschein.metrics.dropwizard;
 
 import com.codahale.metrics.Meter;
 import com.codahale.metrics.MetricRegistry;
