+++ fahrschein-metrics-dropwizard/src/main/java/org/zalando/fahrschein/metrics/dropwizard/LastActivityMetricsCollector.java
@@ -1,4 +1,4 @@
-package org.zalando.fahrschein.metrics;
+package org.zalando.fahrschein.metrics.dropwizard;
 
 import com.codahale.metrics.Gauge;
 import com.codahale.metrics.MetricRegistry;
