+++ fahrschein-metrics-micrometer/src/main/java/org/zalando/fahrschein/metrics/micrometer/LastActivityMetricsCollector.java
@@ -1,4 +1,4 @@
-package org.zalando.fahrschein.metrics;
+package org.zalando.fahrschein.metrics.micrometer;
 
 import io.micrometer.core.instrument.Gauge;
 import io.micrometer.core.instrument.MeterRegistry;
